src/consumer/requirements.txt
oci==2.26.0
kafka-python==2.0.2